NPC beginnings
I added an order console in which accepts orders when you press "e", It spawns an NPC with a text bubble and a list of materials which the order console has access to. This is important as other scripts that need the materials can get them from the order console.