Money Script Clean-up
I forgot to mention in my last description that you can subscribe to the C# event OnMoneyChange when the value of Current changes. Also, you can use ResetMoney() to set current back to its default state. This can be used for restarting the game, which is useful because all the varibles are static.